+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Tweaks to Light Macro
 - AuraEffects support - If AuraEffects by @hi25114 is installed, this will be preferred for creation of actor based aura effects. AuraEffects is the spiritual successor to Active Auras and is a highly performant generic aura management module, with bells and whistles for Aura visualisation. v13 only.
+- Bug that stopped prototype token names updating in v13.
 
 # 6.3.0